Moved iter_equal to test module and modified implementation
Remembered that I only wanted the iter_equal function for testing purposes
so moved it to the test module. Also found that my implementation doesn't
support generators or iterables that do not support the len function,
therefore I basically threw in the towel and just converted them to lists
and directecly compared them. Uses more memory in theory but it works.