Added tests for append method of CircList
Realized that the equality method for CircList returns true if comparing
to an list when that list matches the underlying data structure. I haven't
decided yet if I should keep this functionality and just document it in the
interest of duck-typing, or if I should limit equality to only other
CircList object. I'm leaning towards the later at the moment.Added tests for append method of CircList